Issue #57 mainlined directly.

+++ src/lib/__bi.py
@@ -283,7 +283,7 @@ def id(o):
     }
 
     /* Return object's address as an int on the stack */
-    retval = int_new((int)NATIVE_GET_LOCAL(0), &pr);
+    retval = int_new((intptr_t)NATIVE_GET_LOCAL(0), &pr);
     NATIVE_SET_TOS(pr);
 
     return retval;
